--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/CompiledRegexRunner.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/CompiledRegexRunner.cs
@@ -6,32 +6,21 @@ namespace System.Text.RegularExpressions
 {
     internal sealed class CompiledRegexRunner : RegexRunner
     {
-        private Action<RegexRunner>? _goMethod;
-        private Func<RegexRunner, bool>? _findFirstCharMethod;
-        private Action<RegexRunner>? _initTrackCountMethod;
+        private readonly Action<RegexRunner> _goMethod;
+        private readonly Func<RegexRunner, bool> _findFirstCharMethod;
+        private readonly Action<RegexRunner> _initTrackCountMethod;
 
-        public CompiledRegexRunner() { }
-
-        public void SetDelegates(Action<RegexRunner> go, Func<RegexRunner, bool> firstChar, Action<RegexRunner> trackCount)
+        public CompiledRegexRunner(Action<RegexRunner> go, Func<RegexRunner, bool> firstChar, Action<RegexRunner> trackCount)
         {
             _goMethod = go;
             _findFirstCharMethod = firstChar;
             _initTrackCountMethod = trackCount;
         }
 
-        protected override void Go()
-        {
-            _goMethod!(this);
-        }
+        protected override void Go() => _goMethod(this);
 
-        protected override bool FindFirstChar()
-        {
-            return _findFirstCharMethod!(this);
-        }
+        protected override bool FindFirstChar() => _findFirstCharMethod(this);
 
-        protected override void InitTrackCount()
-        {
-            _initTrackCountMethod!(this);
-        }
+        protected override void InitTrackCount() => _initTrackCountMethod(this);
     }
 }