--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/GroupCollection.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/GroupCollection.cs
@@ -221,6 +221,7 @@ namespace System.Text.RegularExpressions
                 value = null;
                 return false;
             }
+
             value = group;
             return true;
         }