… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCompiler.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexFCD.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexLWCGCompiler.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexNode.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexParser.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexPrefix.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexReplacement.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexRunner.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexWriter.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Ctor.Tests.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Groups.Tests.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Match.Tests.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.MultipleMatches.Tests.cs
… src/libraries/System.Text.RegularExpressions/tests/RegexGroupNameTests.cs