@Tornhoof 2019-11-26
Why is the ctor benchmark quite a bit slower now?