@stephentoub 2019-11-26
> Why is the ctor benchmark quite a bit slower now?

Compiling a Regex gets slower because we're doing more work during compilation, specifically calling CharInClass 128 times to create the ASCII bitmap. But because we do so, we then get to avoid calling CharInClass for every single character we try to match against a character set every time the Regex is used. Compiling a Regex is already an order of magnitude more expensive than not, so you only do it when you're going to be using the Regex over and over and over and over, in which case this is by far worth the trade-off.